removed a lot of warnings

+++ src/Button.cpp
@@ -12,7 +12,7 @@ Button::Button(sf::Vector2f pos, std::string string)
 	box.setPosition(pos);
 	text.setPosition(100, 100);
 	text.setString(string);
-	box.setSize(sf::Vector2f(string.size()*20 + 4, 36));
+	box.setSize(sf::Vector2f((float)(string.size()*20 + 4), (float)36));
 }
 
 void Button::render(sf::RenderWindow* window)
@@ -38,7 +38,7 @@ void Button::setPosition(sf::Vector2f pos)
 void Button::setString(std::string string)
 {
 	text.setString(string);
-	box.setSize(sf::Vector2f(string.size() * 20 + 4, 36));
+	box.setSize(sf::Vector2f((float)(string.size() * 20 + 4), (float)36));
 }
 
 void Button::setFont(sf::Font *font)
